Add middleware and refactor some code
