+++ app.js
@@ -1,6 +1,6 @@
 import express from 'express';
-import db from './db/db';
 import bodyParser from 'body-parser';
+import router from './routes/index';
 
 // Setup express project.
 const app = express();
@@ -9,125 +9,8 @@ const app = express();
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 
-// get all todos
-app.get('/api/v1/todos', (req, res) => {
-    res.status(200).send({
-        success: 'true',
-        message: 'todos retrieved successfully',
-        todos: db
-    })
-});
+app.use(router)
 
-// Create todos
-app.post('/api/v1/todos', (req, res) => {
-    if (!req.body.title) {
-        return res.status(400).send({
-            success: 'false',
-            message: 'Title field is required.'
-
-        });
-    } else if (!req.body.description) {
-        return res.status(400).send({
-            success: 'false',
-            message: 'Description field is required.'
-        });
-    }
-
-    const todo = {
-        id: db.length + 1,
-        title: req.body.title,
-        description: req.body.description
-    }
-
-    db.push(todo);
-
-    return res.status(201).send({
-        success: 'true',
-        message: 'Todo has been successfully created.',
-        todo
-    });
-});
-
-// Get a single todo.
-app.get('/api/v1/todos/:id', (req, res) => {
-    const id = parseInt(req.params.id, 10);
-
-    db.map((todo) => {
-        if (todo.id === id) {
-            return res.status(200).send({
-                success: 'true',
-                message: 'Todo has been retrieved.',
-                todo
-            })
-        }
-    });
-
-    return res.status(404).send({
-        success: 'false',
-        message: 'Todo does not exist'
-    })
-});
-
-// Delete single todo
-app.delete('/api/v1/todos/:id', (req, res) => {
-    const id = parseInt(req.params.id, 10);
-
-    db.map((todo, index) => {
-        if (todo.id === id) {
-            db.splice(index, 1)
-            return res.status(200).send({
-                success: 'true',
-                message: 'Todo has been deleted successfully.'
-            })
-        }
-    });
-
-    return res.status(404).send({
-        success: 'false',
-        message: 'Todo does not exist'
-    })
-});
-
-// Update todo
-app.put('/api/v1/todos/:id', (req, res) => {
-    const id = parseInt(req.params.id, 10);
-    let todoFound;
-    let itemIndex;
-    db.map((todo, index) => {
-        if (todo.id === id) {
-            todoFound = todo;
-            itemIndex = index;
-        }
-    });
-    if (!todoFound) {
-        return res.status(404).send({
-            success: 'false',
-            message: 'todo not found',
-        });
-    }
-    if (!req.body.title) {
-        return res.status(400).send({
-            success: 'false',
-            message: 'title is required',
-        });
-    } else if (!req.body.description) {
-        return res.status(400).send({
-            success: 'false',
-            message: 'description is required',
-        });
-    }
-    const updatedTodo = {
-        id: todoFound.id,
-        title: req.body.title || todoFound.title,
-        description: req.body.description || todoFound.description,
-    };
-    db.splice(itemIndex, 1, updatedTodo);
-    return res.status(201).send({
-        success: 'true',
-        message: 'todo added successfully',
-        updatedTodo,
-    });
-});
 
 const PORT = 5000;
 app.listen(PORT, () => {
